Break i2c_device class into its own file
